Standalone detr.py file for utility code from FB repo
